project_apis/serverless_zappa/requirements.txt
argcomplete==1.12.3
beautifulsoup4==4.10.0
boto3==1.20.16
botocore==1.23.16
certifi==2021.10.8
cfn-flip==1.3.0
charset-normalizer==2.0.8
click==8.0.3
cssselect==1.1.0
Django==2.1.9
djangorestframework==3.10.0
durationpy==0.5
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.0
future==0.18.2
hjson==3.0.2
idna==3.3
jieba3k==0.35.1
jmespath==0.10.0
joblib==1.1.0
kappa==0.6.0
lxml==4.6.4
newspaper3k==0.2.8
nltk==3.6.5
pep517==0.12.0
Pillow==8.4.0
pip-tools==6.4.0
placebo==0.10.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-file==1.5.1
s3transfer==0.5.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
text-unidecode==1.3
textblob==0.17.1
tinysegmenter==0.3
tldextract==3.1.2
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
troposphere==3.1.1
urllib3==1.26.7
vaderSentiment==3.3.2
Werkzeug==2.0.2
whitenoise==5.3.0
wsgi-request-logger==0.4.6
zappa==0.54.1
